Added SliceToDb and SliceToWeb, Tests and updated README.md #3
